requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
anyio==3.7.1
appdirs==1.4.4
arxiv==2.1.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
docker==6.1.3
feedparser==6.0.10
FLAML==2.1.1
fonttools==4.39.4
frozendict==2.3.8
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
gpt-index==0.4.24
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langsmith==0.0.75
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdurl==0.1.2
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.26.1
openai==1.6.1
openapi-schema-pydantic==1.2.4
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
protobuf==4.24.4
pyarrow==13.0.0
pyautogen==0.2.2
pycparser==2.21
pydantic==1.10.8
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
selenium==4.16.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
streamlit==1.28.0
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.6.4
wsproto==1.2.0
yarl==1.9.2
yfinance==0.2.18
zipp==3.17.0
